02_TOOLSETS/rail_cost_model/src/main.py
# Main entry point
import argparse
import logging
from pathlib import Path
import yaml

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

if __name__ == "__main__":
    print("Rail Cost Model CLI")
